+++ packages/components/src/components/action/button-pure/button-pure.tsx
@@ -6,8 +6,7 @@ import {
   improveButtonHandlingForCustomElement,
   improveFocusHandlingForCustomElement,
   isDark,
-  mapBreakpointPropToPrefixedClasses,
-  prefix,
+  mapBreakpointPropToClasses,
   transitionListener,
 } from '../../../utils';
 import type { BreakpointCustomizable, ButtonType, IconName, TextSize, TextWeight, Theme } from '../../../types';
@@ -29,7 +28,7 @@ export class ButtonPure {
   @Prop() public type?: ButtonType = 'submit';
 
   /** Disables the button. No events will be triggered while disabled state is active. */
-  @Prop({ reflect: true }) public disabled?: boolean = false;
+  @Prop() public disabled?: boolean = false;
 
   /** Disables the button and shows a loading indicator. No events will be triggered while loading state is active. */
   @Prop() public loading?: boolean = false;
@@ -81,28 +80,16 @@ export class ButtonPure {
 
   public render(): JSX.Element {
     const buttonPureClasses = {
-      [prefix('button-pure')]: true,
-      [prefix('button-pure--theme-dark')]: isDark(this.theme),
-      ...mapBreakpointPropToPrefixedClasses('button-pure--size', this.size),
-    };
-
-    const labelClasses = {
-      [prefix('button-pure__label')]: true,
-      ...mapBreakpointPropToPrefixedClasses('button-pure__label-', this.hideLabel, {
-        classSuffixes: ['hidden', 'visible'],
-      }),
-    };
-
-    const sublineClasses = {
-      [prefix('button-pure__subline')]: true,
-      ...mapBreakpointPropToPrefixedClasses('button-pure__subline-', this.hideLabel, {
-        classSuffixes: ['hidden', 'visible'],
-      }),
+      ['root']: true,
+      ['root--theme-dark']: isDark(this.theme),
+      ...mapBreakpointPropToClasses('root--size', this.size),
+      ...mapBreakpointPropToClasses('root-', this.hideLabel, ['without-label', 'with-label']),
     };
 
     const iconProps = {
-      class: prefix('button-pure__icon'),
+      class: 'icon',
       size: 'inherit',
+      theme: this.theme,
       ref: (el: HTMLElement) => (this.iconTag = el),
     };
 
@@ -119,7 +106,7 @@ export class ButtonPure {
           aria-busy={this.loading ? 'true' : null}
         >
           {this.loading ? (
-            <PrefixedTagNames.pSpinner {...iconProps} theme={this.theme} />
+            <PrefixedTagNames.pSpinner {...iconProps} />
           ) : (
             <PrefixedTagNames.pIcon
               {...iconProps}
@@ -129,12 +116,12 @@ export class ButtonPure {
               aria-hidden="true"
             />
           )}
-          <PrefixedTagNames.pText class={labelClasses} tag="span" color="inherit" size="inherit" weight={this.weight}>
+          <PrefixedTagNames.pText class="label" tag="span" color="inherit" size="inherit" weight={this.weight}>
             <slot />
           </PrefixedTagNames.pText>
         </button>
         {hasNamedSlot(this.host, 'subline') && (
-          <PrefixedTagNames.pText class={sublineClasses} tag="div" color="inherit" size="inherit">
+          <PrefixedTagNames.pText class="subline" tag="div" color="inherit" size="inherit">
             <slot name="subline" />
           </PrefixedTagNames.pText>
         )}