+++ packages/components/src/components/content/modal/modal.tsx
@@ -1,7 +1,13 @@
 import { Component, Event, EventEmitter, Element, h, JSX, Prop, Watch, Host } from '@stencil/core';
 import type { BreakpointCustomizable } from '../../../types';
-import { getPrefixedTagNames, mapBreakpointPropToPrefixedClasses, prefix } from '../../../utils';
-import { getFirstAndLastElement, getFocusableElements, getScrollTopOnTouch, setScrollLock } from './modal-utils';
+import { getPrefixedTagNames, mapBreakpointPropToClasses } from '../../../utils';
+import {
+  addCss,
+  getFirstAndLastElement,
+  getFocusableElements,
+  getScrollTopOnTouch,
+  setScrollLock,
+} from './modal-utils';
 
 @Component({
   tag: 'p-modal',
@@ -12,7 +18,7 @@ export class Modal {
   @Element() public host!: HTMLElement;
 
   /** If true, the modal is open. */
-  @Prop({ reflect: true }) public open = false;
+  @Prop() public open = false;
   /** If true, the modal will not have a close button. */
   @Prop() public disableCloseButton?: boolean = false;
   /** If true, the modal will not be closable via backdrop click. */
@@ -36,7 +42,6 @@ export class Modal {
     if (isOpen) {
       this.focusableElements = getFocusableElements(this.host, this.closeBtn);
       this.focusedElBeforeOpen = document.activeElement as HTMLElement;
-      this.focusableElements[0]?.focus();
     } else {
       this.focusedElBeforeOpen?.focus();
     }
@@ -55,6 +60,18 @@ export class Modal {
     this.focusableElements = getFocusableElements(this.host, this.closeBtn);
   }
 
+  public componentWillRender(): void {
+    addCss(this.host, this.open);
+  }
+
+  public componentDidUpdate(): void {
+    if (this.open) {
+      /* the close button is not immediately visible when the  @Watch('open') triggers,
+       so we focus it in componentDidUpdate() */
+      this.focusableElements[0]?.focus();
+    }
+  }
+
   public disconnectedCallback(): void {
     this.setKeyboardListener(false);
     setScrollLock(this.host, false, this.setScrollTop);
@@ -63,15 +80,9 @@ export class Modal {
   public render(): JSX.Element {
     const hasHeader = this.heading || !this.disableCloseButton;
     const rootClasses = {
-      [prefix('modal')]: true,
-      ...mapBreakpointPropToPrefixedClasses('modal-', this.fullscreen, {
-        classSuffixes: ['fullscreen-on', 'fullscreen-off'],
-      }),
+      ['root']: true,
+      ...mapBreakpointPropToClasses('root-', this.fullscreen, ['fullscreen-on', 'fullscreen-off']),
     };
-    const headerClasses = prefix('modal__header');
-    const btnCloseWrapperClasses = prefix('modal__close');
-    const btnCloseClasses = prefix('modal__close-button');
-
     const PrefixedTagNames = getPrefixedTagNames(this.host);
 
     return (
@@ -84,16 +95,15 @@ export class Modal {
           aria-hidden={!this.open ? 'true' : 'false'}
         >
           {hasHeader && (
-            <header class={headerClasses}>
+            <header>
               {this.heading && (
                 <PrefixedTagNames.pHeadline variant={{ base: 'medium', m: 'large' }}>
                   {this.heading}
                 </PrefixedTagNames.pHeadline>
               )}
               {!this.disableCloseButton && (
-                <div class={btnCloseWrapperClasses}>
+                <div class="close">
                   <PrefixedTagNames.pButtonPure
-                    class={btnCloseClasses}
                     ref={(el) => (this.closeBtn = el)}
                     hideLabel
                     icon="close"