+++ packages/components/src/components/navigation/link/link.tsx
@@ -5,8 +5,7 @@ import {
   improveFocusHandlingForCustomElement,
   insertSlottedStyles,
   isDark,
-  mapBreakpointPropToPrefixedClasses,
-  prefix,
+  mapBreakpointPropToClasses,
 } from '../../../utils';
 import type { BreakpointCustomizable, IconName, LinkTarget, LinkVariant, Theme } from '../../../types';
 
@@ -54,15 +53,11 @@ export class Link {
     const TagType = this.href === undefined ? 'span' : 'a';
 
     const linkClasses = {
-      [prefix('link')]: true,
-      [prefix(`link--${this.variant}`)]: true,
-      [prefix('link--theme-dark')]: isDark(this.theme),
-      ...mapBreakpointPropToPrefixedClasses('link-', this.hideLabel, {
-        classSuffixes: ['without-label', 'with-label'],
-      }),
+      ['root']: true,
+      [`root--${this.variant}`]: this.variant !== 'secondary',
+      ['root--theme-dark']: isDark(this.theme),
+      ...mapBreakpointPropToClasses('root-', this.hideLabel, ['without-label', 'with-label']),
     };
-    const iconClasses = prefix('link__icon');
-    const labelClasses = prefix('link__label');
 
     const PrefixedTagNames = getPrefixedTagNames(this.host);
 
@@ -77,14 +72,14 @@ export class Link {
         })}
       >
         <PrefixedTagNames.pIcon
-          class={iconClasses}
+          class="icon"
           size="inherit"
           name={this.icon}
           source={this.iconSource}
           color="inherit"
           aria-hidden="true"
         />
-        <PrefixedTagNames.pText tag="span" color="inherit" class={labelClasses}>
+        <PrefixedTagNames.pText tag="span" color="inherit" class="label">
           <slot />
         </PrefixedTagNames.pText>
       </TagType>