+++ packages/components/src/utils/breakpoint-customizable.ts
@@ -11,14 +11,13 @@ export type BreakpointValues<T> = {
 // string is needed in order to pass and parse objects via prop decorator
 export type BreakpointCustomizable<T> = T | BreakpointValues<T> | string;
 
-type BreakpointValue = string | number | boolean; // TODO: replace with generic T
+type BreakpointValue = string | number | boolean;
 type ClassSuffixes = [string, string];
 
 type JSXClasses = {
   [className: string]: boolean;
 };
 
-/* eslint-disable @typescript-eslint/indent */
 export const parseJSON = (
   prop: BreakpointCustomizable<BreakpointValue>
 ): BreakpointValues<BreakpointValue> | BreakpointValue => {
@@ -41,53 +40,79 @@ export const parseJSON = (
 };
 
 const getClassName = (value: BreakpointValue, classSuffixes: ClassSuffixes): string =>
-  typeof value === 'boolean' ? classSuffixes[value ? 0 : 1] : value.toString();
+  typeof value === 'boolean' ? classSuffixes[value ? 0 : 1] : `${value}`;
 
 const getBreakpointSuffix = (breakpoint: Breakpoint): string => (breakpoint !== 'base' ? `-${breakpoint}` : '');
 
+/**
+ * @deprecated
+ */
 const createClass = (
   classPrefix: string,
   value: BreakpointValue,
   breakpoint: Breakpoint,
-  classSuffixes: ClassSuffixes,
-  disablePrefixP?: boolean
+  classSuffixes: ClassSuffixes
 ): JSXClasses => {
   if (value !== null && value !== undefined) {
     const className = `${classPrefix}-${getClassName(value, classSuffixes)}${getBreakpointSuffix(breakpoint)}`;
     return {
-      [disablePrefixP ? className : prefix(className)]: true,
+      [prefix(className)]: true,
     };
   }
 };
 
 /**
+ * @deprecated
  * @param classPrefix
  * @param prop
  * @param classSuffixes
  *  First value in array is used for true boolean values, second for false.
  *  It's only used when prop is a boolean or prop is a object/JSON5 that contains a boolean.
  */
-type Options = {
-  classSuffixes?: ClassSuffixes;
-  disablePrefixP?: boolean;
-};
-
 export const mapBreakpointPropToPrefixedClasses = (
   classPrefix: string,
   prop: BreakpointCustomizable<BreakpointValue>,
-  opts?: Options
+  classSuffixes?: ClassSuffixes
 ): JSXClasses => {
   const parsedProp = parseJSON(prop);
 
-  const { classSuffixes, disablePrefixP } = opts ?? {};
+  return typeof parsedProp === 'object'
+    ? Object.entries(parsedProp).reduce(
+        (classes, [breakpoint, value]: [Breakpoint, BreakpointValue]) => ({
+          ...classes,
+          ...createClass(classPrefix, value, breakpoint, classSuffixes),
+        }),
+        {} as JSXClasses
+      )
+    : createClass(classPrefix, parsedProp, 'base', classSuffixes);
+};
+
+const createJSXClass = (
+  classPrefix: string,
+  value: BreakpointValue,
+  breakpoint: Breakpoint,
+  classSuffixes: ClassSuffixes
+): JSXClasses => {
+  if (value !== null && value !== undefined) {
+    const className = `${classPrefix}-${getClassName(value, classSuffixes)}${getBreakpointSuffix(breakpoint)}`;
+    return { [className]: true };
+  }
+};
+
+export const mapBreakpointPropToClasses = (
+  classPrefix: string,
+  prop: BreakpointCustomizable<BreakpointValue>,
+  classSuffixes?: ClassSuffixes
+): JSXClasses => {
+  const parsedProp = parseJSON(prop);
 
   return typeof parsedProp === 'object'
     ? Object.entries(parsedProp).reduce(
-        (classes, [breakpoint, value]) => ({
+        (classes, [breakpoint, value]: [Breakpoint, BreakpointValue]) => ({
           ...classes,
-          ...createClass(classPrefix, value, breakpoint as Breakpoint, classSuffixes, disablePrefixP),
+          ...createJSXClass(classPrefix, value, breakpoint, classSuffixes),
         }),
-        {}
+        {} as JSXClasses
       )
-    : createClass(classPrefix, parsedProp, 'base', classSuffixes, disablePrefixP);
+    : createJSXClass(classPrefix, parsedProp, 'base', classSuffixes);
 };